==> code/3.statistical_analysis_compare_T1_T2.Rmd <==
---
title: "DMD mdx corticosteroids project: Statistical analysis - Compare top peaks from genotype comparison at T0 at T1-T2"
author: Simina Boca
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path = "../figures/")
```

```{r}
library(dplyr)
library(reshape2)
library(ggplot2)
library(nlme)
library(qvalue)
library(limma) #for vennDiagram function
source("functions.R")

```

## Load data, previous analysis results

Read in data, load results from genotype comparison at time T0:

```{r}
##load analysis results
load("../objs/stat_analysis_geno_T0.RData")

##clarify what the p-values are from
p_vals_T0 <- p_vals
##read in formatted data
MasterFrame <- read.delim(file="../data/Supplemental Table 1.tsv",
                          header=TRUE,
                          sep="\t",
                          check.names=FALSE)
```

## Prepare data 

Prepare data for analysis:

```{r}
##first take out the QC samples
MasterFrameNoQC <- MasterFrame[MasterFrame$SampleID != "QC",]
##only include interaction terms between Times T1/T2 and Group variables (since just one group 2 at baseline)
MasterFrameNoQC$TimeT1 <- MasterFrameNoQC$TimeT2 <- 
  MasterFrameNoQC$GenoMDX <- 
  MasterFrameNoQC$GroupPred <- 
  rep(0, nrow(MasterFrameNoQC)) 
MasterFrameNoQC$TimeT1[MasterFrameNoQC$Time=="T1"] <- 1
MasterFrameNoQC$TimeT2[MasterFrameNoQC$Time=="T2"] <- 1
MasterFrameNoQC$GenoMDX[MasterFrameNoQC$Genotype=="mdx23"] <- 1
MasterFrameNoQC$GroupPred[MasterFrameNoQC$Group=="group 1"] <- 1

##get metabolites
mets <- colnames(MasterFrameNoQC)[grep("m/z",colnames(MasterFrameNoQC))]

MasterFrame_T0 <- subset(MasterFrameNoQC, Time=="T0")
##remove any metabolites with variance = 0
var0 <- names(which(apply(MasterFrame_T0[mets],2,var)==0))
var0
MasterFrame_T0 <- dplyr::select(MasterFrame_T0, -one_of(var0))
mets <- setdiff(mets, var0)

length(mets)

##now only keep these metabolites for the T1 vs T2 comparison ie remove the variance 0 ones
MasterFrameNoQC <- dplyr::select(MasterFrameNoQC, -one_of(var0))
##only keep times T1 and T2
MasterFrame_T1_T2 <- subset(MasterFrameNoQC, Time!="T0")
MasterFrame_T1_T2$Time <- as.character(MasterFrame_T1_T2$Time)

all_mets <- mets
```

## Get some descriptive tables

Time T0:
```{r}
table(MasterFrame_T0$Genotype)
```

Time T1:
```{r}
table(MasterFrame_T1_T2$Group[MasterFrame_T1_T2$Time=="T1"],
      MasterFrame_T1_T2$Genotype[MasterFrame_T1_T2$Time=="T1"])
```

Time T2:
```{r}
table(MasterFrame_T1_T2$Group[MasterFrame_T1_T2$Time=="T2"],
      MasterFrame_T1_T2$Genotype[MasterFrame_T1_T2$Time=="T2"])
```

Overlap between samples at Time 1 and Time 2:
```{r}
samples_times <- table(MasterFrame_T1_T2$SampleID,
                       MasterFrame_T1_T2$Time)
length(unique(MasterFrame_T1_T2$SampleID))
colSums(samples_times)
sum(samples_times[,1]*(1-samples_times[,2]))
sum(samples_times[,2]*(1-samples_times[,1]))
```

## Make plots

First label the interesting metabolites:
```{r}
met_top <- "m/z = 357.24594, mode = Positive, RT = 20.659"
met_creatine <- "m/z = 132.07757, mode = Positive, RT = 23.18143"
met_creatinine <- "m/z = 114.06447, mode = Positive, RT = 22.59731"
met_hypusine <- "m/z = 234.18087, mode = Positive, RT = 19.39639"
met_biliverdin <- "m/z = 583.25604, mode = Positive, RT = 381.06653"
```

Change genotype from "mdx23" to "mdx":
```{r}
MasterFrame_T1_T2$Genotype <- as.character(MasterFrame_T1_T2$Genotype)
MasterFrame_T1_T2$Genotype[MasterFrame_T1_T2$Genotype == "mdx23"] <- "mdx"

MasterFrame_T1_T2$Treatment <- MasterFrame_T1_T2$Group
MasterFrame_T1_T2$Treatment <- as.character(MasterFrame_T1_T2$Treatment)
MasterFrame_T1_T2$Treatment[MasterFrame_T1_T2$Treatment == "group 1"] <- "prednisolone"
MasterFrame_T1_T2$Treatment[MasterFrame_T1_T2$Treatment == "group 2"] <- "vehicle control"

```

```{r plots_T1_T2_top_row, fig.height=3, fig.width=15}
plot1 <- ggplot(MasterFrame_T1_T2, aes_string(x="Time", y=paste("`",met_top,"`",sep=""),
                                              group="SampleID", color="Genotype", shape="Genotype", 
                                              linetype="Treatment")) +
  geom_point() +
  geom_line() +
  scale_color_manual(name = "Genotype", 
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(2, 4)) +
  scale_shape_manual(name="Genotype",
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(17, 16)) +
  ylab("Normalized, log2-transformed value") + 
  ggtitle("a)\nm/z = 357.25, Positive mode") +
  theme_bw() +
  theme(plot.title = element_text(size = 13, hjust=0.5),
        axis.title = element_text(size=12),
        axis.text = element_text(size=11),
        legend.text = element_text(size=11))

plot2 <- ggplot(MasterFrame_T1_T2, aes_string(x="Time", y=paste("`",met_creatinine,"`",sep=""),
                                              group="SampleID", color="Genotype", shape="Genotype", 
                                              linetype="Treatment")) +
  geom_point() +
  geom_line() +
  scale_color_manual(name = "Genotype", 
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(2, 4)) +
  scale_shape_manual(name="Genotype",
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(17, 16)) +
  ylab("Normalized, log2-transformed value") + 
  ggtitle("b)\nCreatinine") +
  theme_bw() +
  theme(plot.title = element_text(size = 13, hjust=0.5),
        axis.title = element_text(size=12),
        axis.text = element_text(size=11),
        legend.text = element_text(size=11))

plot3 <- ggplot(MasterFrame_T1_T2, aes_string(x="Time", y=paste("`",met_creatine,"`",sep=""),
                                              group="SampleID", color="Genotype", shape="Genotype", 
                                              linetype="Treatment")) +
  geom_point() +
  geom_line() +
  scale_color_manual(name = "Genotype", 
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(2, 4)) +
  scale_shape_manual(name="Genotype",
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(17, 16)) +
  ylab("Normalized, log2-transformed value") + 
  ggtitle("c)\nCreatine") +
  theme_bw() +
  theme(plot.title = element_text(size = 13, hjust=0.5),
        axis.title = element_text(size=12),
        axis.text = element_text(size=11),
        legend.text = element_text(size=11))

multiplot(plot1, plot2, plot3, cols=3)
```

```{r plots_T1_T2_bottom_row, fig.height=3, fig.width=10}
plot4 <- ggplot(MasterFrame_T1_T2, aes_string(x="Time", y=paste("`",met_hypusine,"`",sep=""),
                                              group="SampleID", color="Genotype", shape="Genotype", 
                                              linetype="Treatment")) +
  geom_point() +
  geom_line() +
  scale_color_manual(name = "Genotype", 
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(2, 4)) +
  scale_shape_manual(name="Genotype",
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(17, 16)) +
  ylab("Normalized, log2-transformed value") + 
  ggtitle("d)\nHypusine") +
  theme_bw() +
  theme(plot.title = element_text(size = 13, hjust=0.5),
        axis.title = element_text(size=12),
        axis.text = element_text(size=11),
        legend.text = element_text(size=11))

plot5 <- ggplot(MasterFrame_T1_T2, aes_string(x="Time", y=paste("`",met_biliverdin,"`",sep=""),
                                              group="SampleID", color="Genotype", shape="Genotype", 
                                              linetype="Treatment")) +
  geom_point() +
  geom_line() +
  scale_color_manual(name = "Genotype", 
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(2, 4)) +
  scale_shape_manual(name="Genotype",
                     breaks = c("WT", "mdx"),
                     labels = c("WT", "mdx"),
                     values = c(17, 16)) +
  ylab("Normalized, log2-transformed value") + 
  ggtitle("e)\nBiliverdin") +
  theme_bw() +
  theme(plot.title = element_text(size = 13, hjust=0.5),
        axis.title = element_text(size=12),
        axis.text = element_text(size=11),
        legend.text = element_text(size=11))

multiplot(plot4, plot5, cols=2)
```

Compare values of creatine and biliverdin at T0 for WT for Group 1 vs Group 2.
Are they significantly different from each other? Also do this for mdx separately.

```{r}
MasterFrame_T0_WT <- filter(MasterFrame_T0, GenoMDX==0)
MasterFrame_T0_mdx <- filter(MasterFrame_T0, GenoMDX==1)

t.test(MasterFrame_T0_WT[MasterFrame_T0_WT$Group=="group 1",met_creatine],
       MasterFrame_T0_WT[MasterFrame_T0_WT$Group=="group 2",met_creatine],
       var.equal = TRUE)
t.test(MasterFrame_T0_mdx[MasterFrame_T0_WT$Group=="group 1",met_creatine],
       MasterFrame_T0_mdx[MasterFrame_T0_WT$Group=="group 2",met_creatine],
       var.equal = TRUE)

t.test(MasterFrame_T0_WT[MasterFrame_T0_WT$Group=="group 1",met_biliverdin],
       MasterFrame_T0_WT[MasterFrame_T0_WT$Group=="group 2",met_biliverdin],
       var.equal = TRUE)
t.test(MasterFrame_T0_mdx[MasterFrame_T0_WT$Group=="group 1",met_biliverdin],
       MasterFrame_T0_mdx[MasterFrame_T0_WT$Group=="group 2",met_biliverdin],
       var.equal = TRUE)
```

## Test hypotheses relating to impact of genotype, group, and time for the top 50 metabolites from the T0 analysis + creatine, creatinine

Fit models (include creatine and creatinine here again):

```{r}
top_peaks <- names(which(rank(sort(p_vals_T0))<=50))
top_peaks <- c(top_peaks, met_creatine, met_creatinine)

##full models
lme_full <- lapply(MasterFrame_T1_T2[,top_peaks],
                    function(metInt, geno, group, Times, ID){
                      ##full model
                      lme(metInt ~ geno*Times*group, random =~ 1 | ID,
                          method="ML");
                    }, 
                    MasterFrame_T1_T2$GenoMDX, MasterFrame_T1_T2$GroupPred, 
                    MasterFrame_T1_T2$Time,
                    MasterFrame_T1_T2$SampleID)
##models with no genotypes
lme_no_geno <- lapply(MasterFrame_T1_T2[,top_peaks],
                      function(metInt, group, Times, ID){
                        ##full model
                        lme(metInt ~ Times*group, random =~ 1 | ID,
                            method="ML");
                      }, 
                      MasterFrame_T1_T2$GroupPred, 
                      MasterFrame_T1_T2$Time,
                      MasterFrame_T1_T2$SampleID)
##models with no group
lme_no_group <- lapply(MasterFrame_T1_T2[,top_peaks],
                      function(metInt, geno, Times, ID){
                        ##full model
                        lme(metInt ~ Times*geno, random =~ 1 | ID,
                            method="ML");
                      }, 
                      MasterFrame_T1_T2$GenoMDX, 
                      MasterFrame_T1_T2$Time,
                      MasterFrame_T1_T2$SampleID)
##models with no time
lme_no_time <- lapply(MasterFrame_T1_T2[,top_peaks],
                      function(metInt, geno, group, ID){
                        ##full model
                        lme(metInt ~ geno*group, random =~ 1 | ID,
                            method="ML");
                      }, 
                      MasterFrame_T1_T2$GenoMDX, 
                      MasterFrame_T1_T2$GroupPred,
                      MasterFrame_T1_T2$SampleID)


```

Used saved models to perform association tests using ANOVA:

```{r}
length(all_mets)

##associations with genotype
p_geno <- mapply(function(mod1,mod2){anova(mod1,mod2)["p-value"][2,1]},
                 lme_full, lme_no_geno)
top_assoc_geno <- names(which(p_geno*length(all_mets) < 0.05))
length(top_assoc_geno)
p_geno[top_assoc_geno]
setdiff(top_peaks, top_assoc_geno)

##associations with group
p_group <- mapply(function(mod1,mod2){anova(mod1,mod2)["p-value"][2,1]},
                 lme_full, lme_no_group)
p_group <- sort(p_group)
##get the top metabolites in a vector
top_assoc_group <- names(which(sort(p_group)*length(all_mets) < 0.05))
length(top_assoc_group)
p_group[top_assoc_group]
##also look at other interesting metabolites here
p_group[met_top]
p_group[met_hypusine]

##associations with time
p_time <- mapply(function(mod1,mod2){anova(mod1,mod2)["p-value"][2,1]},
                 lme_full, lme_no_time)
##get the top metabolites in a vector
top_assoc_time <- names(which(sort(p_time)*length(all_mets) < 0.05))
length(top_assoc_time)
p_time[top_assoc_time]

##get which ones are associated with group but not time and vice-versa
setdiff(top_assoc_group, top_assoc_time)
setdiff(top_assoc_time, top_assoc_group)
##what about just among the top 10 associations with time?
setdiff(top_assoc_time[1:10], top_assoc_group)

```

Do a Venn diagram:

```{r venn_significant_overlaps, fig.height=5, fig.width=7}
all_peaks <- data.frame(matrix(0, nrow=length(top_peaks),
                               ncol=4))
colnames(all_peaks) <- c("Genotype (T0)",
                         "Treatment",
                         "Time",
                         "Genotype")
rownames(all_peaks) <- top_peaks
all_peaks[,1] <- 1
all_peaks[top_assoc_group,2] <- 1
all_peaks[top_assoc_time,3] <- 1
all_peaks[top_assoc_geno,4] <- 1

# v <- vennCounts(all_peaks)
# 
# vennDiagram(v)

v2 <- vennCounts(all_peaks[,2:4])

vennDiagram(v2,
            circle.col = c("orange",
                           "skyblue",
                           "pink1"))
```

Save results in flat file:
```{r}
lme_full_results_table <- 
  data.frame()

for(metab in names(lme_full))
{
  ##get fixed effects info
  summary_coef_metab <- coef(summary(lme_full[[metab]]))
  
  ##get random effects estimated standard deviations
  random_metab_SD <- as.numeric(as.character(unlist(VarCorr(lme_full[[metab]])[1,"StdDev"])))
  
  ##get p-value from the 3 likelihood ratio tests (from the anova function)
  p_vals_LRT <- c(p_geno[metab], p_time[metab], p_group[metab])
  
  ##get 95% confidence intervals
  CIs <- intervals(lme_full[[metab]], level=0.95, which="fixed")$fixed
  
  ##first arrange this in proper way:
  summary_results_metab <- c(summary_coef_metab[,"Value"],
                             summary_coef_metab[,"Std.Error"],
                             summary_coef_metab[,"DF"],
                             CIs[,"lower"],
                             CIs[,"upper"],
                             summary_coef_metab[,"t-value"],
                             summary_coef_metab[,"p-value"],
                             random_metab_SD,
                             p_vals_LRT)
  summary_results_metab <- data.frame(matrix(summary_results_metab,nrow=1))
  summary_results_metab <- cbind(metab, summary_results_metab)
  ##now dump this into table with all metabolites
  lme_full_results_table <- rbind(lme_full_results_table,
                                  summary_results_metab)
}
##fix column names
##preprend letters to make the sorting easier
colnames(lme_full_results_table)[1] <- "(AA).Metabolite"
colnames(lme_full_results_table)[-1] <- 
  c(paste(rownames(summary_coef_metab),
          rep(c("A.Beta","B.SE","C.DF",
                "D.95 CI lower", "E.95 CI upper",
                "F.T","G.P"),
              each=nrow(summary_coef_metab))),
    "YY.random intercept SD",
    "ZZ.P LRT genotype",
    "ZZ.P LRT time",
    "ZZ.P LRT treatment")

##capitalize everything, change "group" to "treatment"
colnames(lme_full_results_table) <-
  toupper(gsub("group","treatment",
               colnames(lme_full_results_table)))

lme_full_results_table <- lme_full_results_table[,order(colnames(lme_full_results_table))]
                                
##now take out the preprended stuff
colnames(lme_full_results_table) <-
  gsub("\\(AA\\)\\.|A\\.|B\\.|C\\.|D\\.|E\\.|F\\.|G\\.|YY\\.|ZZ\\.","",
       colnames(lme_full_results_table))

##now save to CSV file
write.csv(lme_full_results_table, file="../lme_models_times_T1_T2_summary_results.csv",
          row.names = FALSE, quote = FALSE)

```

## Now make some plots for the top metabolites associated with group

```{r, top_T0_geno_assoc_group_T1_T2}
MasterFrame_T0_WT <- filter(MasterFrame_T0, GenoMDX==0)
MasterFrame_T0_MDX <- filter(MasterFrame_T0, GenoMDX==1)

##look specifically at these 2 metabolites:
met_484 <- "m/z = 484.16091, mode = Negative, RT = 155.62501"
met_885 <- "m/z = 884.92954, mode = Positive, RT = 231.26267"
##double-check that they are among the top ones with group
setdiff(c(met_484, met_885), top_assoc_group)

for(met in c(met_484, met_885))
{  
  if(met == met_484)
  {
    met_title <- "m/z = 484.16, Negative mode"
  }
  if(met == met_885)
  {
    met_title <- "m/z = 884.93, Positive mode"
  }
  
  g3 <- ggplot(MasterFrame_T1_T2, aes_string(x="Time", y=paste("`",met,"`",sep=""),
                                             group="SampleID", color="Genotype", shape="Genotype", 
                                             linetype="Treatment")) +
    geom_point() + geom_line() +
    ggtitle(met_title) +
    scale_color_manual(name = "Genotype", 
                       breaks = c("WT", "mdx"),
                       labels = c("WT", "mdx"),
                       values = c(2, 4)) +
    scale_shape_manual(name="Genotype",
                       breaks = c("WT", "mdx"),
                       labels = c("WT", "mdx"),
                       values = c(17, 16)) +
    ylab("Normalized, log2-transformed values") + 
    theme_bw() +
    theme(plot.title = element_text(size = 13, hjust=0.5),
          axis.title = element_text(size=12),
          axis.text = element_text(size=11),
          legend.text = element_text(size=11))
  
  print(g3)
}  

```

## Session info

```{r}
sessionInfo()
```
